Moved history migration job to migrate_history.py
