+++ backend/util/expire_orders.py
@@ -24,13 +24,10 @@ MIGRATION_SLEEP_TIME = int(os.getenv("MIGRATION_LOOP_DELAY", "0"))
 async def lifespan(_: FastAPI):
     """Runs before and after the server starts and closes"""
     thread = ExpireOrdersThread()
-    migration_thread = MigrateOrderHistoryThread()
 
     thread.start()
-    migration_thread.start()
     yield  # Anything after here runs when server is shutting down
     thread.stop()
-    migration_thread.stop()
 
 
 class ExpireOrdersThread(threading.Thread):
@@ -59,30 +56,3 @@ class ExpireOrdersThread(threading.Thread):
 
             self._stop_event.wait(SLEEP_TIME)
 
-class MigrateOrderHistoryThread(threading.Thread):
-    """Infinitely running thread to migrate orders."""
-    
-    def __init__ (self):
-        super().__init__()
-        self._stop_event = threading.Event()
-
-    def stop(self):
-        super().__init__()
-        self._stop_event.set()
-
-    def run(self):
-        """"""
-
-        if MIGRATION_SLEEP_TIME  == 0:
-            return
-
-        last_migrated_hour = None
-
-        
-
-        while not self._stop_event.is_set():
-            current_hour = datetime.now().replace(minute=0, second=0, microsecond=0)
-            if last_migrated_hour != current_hour:
-                supabase_middleman.migrate_price_changes(current_hour)
-                last_migrated_hour = current_hour
-            self._stop_event.wait(MIGRATION_SLEEP_TIME)
